backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
ruff>=0.8.0
